Expose repl task dependencies
The repl task is special because it can't run in a pod. It needs to
run in the project context and have access to the environment of the
build.boot script.

In order to keep the project classpath pristine when the REPL is not in
use, the nREPL dependencies are not loaded until the repl task actually
needs to run. This presents difficulties with middleware, since most of
these are in namespaces that assume tools.nrepl is available.

To support middlware in this environment we expose two atoms:

- boot.repl/*default-dependencies* atom containing a vector of maven
  coordinates in the (set-env! :dependencies '[...]) format. These
  dependencies will be added only when the repl task is run, and only if
  the project does not already have explicit dependencies for the deps
  it would otherwise load.

- boot.repl/*default-middleware* atom containing a vector of namespace
  qualified symbols corresponding to desired middleware. The repl task
  will resolve them at runtime as necessary, so they don't need to be
  resolvable from the build.boot.

Modify these to change dependencies or middleware loaded by default by
the repl task. The middleware option to the repl task adds middleware in
addition to these defaults. The handler option will override these.