fix relative-to's accidental quoting of special characters
relative-to used to convert paths to URIs, which resulted in some
special characters being quoted.

This led to errors like the following when trying to start a repl:
,----
|             clojure.lang.ExceptionInfo: java.io.FileNotFoundException: /home/ralf/.boot/tmp/home/ralf/boot-sandbox/dsn/-x24pa9/%23core.clj%23 (No such file or directory)
|     data: {:file "../tmp/boot.user3335412700912634473.clj", :line 41}
| java.util.concurrent.ExecutionException: java.io.FileNotFoundException: /home/ralf/.boot/tmp/home/ralf/boot-sandbox/dsn/-x24pa9/%23core.clj%23 (No such file or directory)
|           java.io.FileNotFoundException: /home/ralf/.boot/tmp/home/ralf/boot-sandbox/dsn/-x24pa9/%23core.clj%23 (No such file or directory)
`----

Here, relative-to quoted an emacs created backup file. So, this may be
someone's master plan to convert people to vim.

Anyway, we can now use JDK 7's nio Path class in order to relativize
Paths.

Note that nio's Path class now throws an error when called with a
relative and an absolute Path. The old relative-to returned a wrong
result in that case. So, this change may reveal more bugs...