Update README.md
The goal is to explain the difference between the version of the wrapper
and the "real Boot," i.e., answer the "why is the latest version of Boot
available for download 2.0, while I can see the actual latest is 2.1.2"
questions.

Should fix https://github.com/boot-clj/boot/issues/238 and the likes of
it.