Add unpeeling of ExecutionException in boot.core/boot and fixes #405
After discussion on Slack it emerged that it is necessary and robust to unpeel ExecutionExceptions
when dereffing a feature so that the App.Exit (or whatever) inside can be properly propagated.

After tracking down all the @ and deref, I came to the conclusion that only boot.core/boot, the
entry point of the execution, needs this.

This patch does exactly that.