Include contents of symlinked directories in fileset
Symlinked directories present in e.g. source-paths were previously
ignored when building the fileset because the relevant tree-walking
functions ignored links. By enabling them to follow symlinks, the
contents of linked directories are now included in the fileset.

Note that the symlinked directories are synced into "real" directories
in the temp dir. For example, a symlink `foo` that points to a directory
containing `bar.clj`, will be present in the temp dir as a *plain*
directory `foo` containing `bar.clj`. So once these symlinked
directories have been added to the fileset, they behave exactly as a
plain directory would (because they are one). As a bonus, this means the
change has no effect on any further temp dir related filesystem interaction
(like scary deletion/cleanup that could interact with symlinks).

Fixes #481.