Prevent play! from throwing on systems with no audio output
It's possible that we could detect the presence/absence of an audio
device using the Java Sound API AudioSystem, but it's not clear to me
that would be more robust.  This will also change problems like trying to
play an supported format, or passing a non-existent file name into a
printed warning instead of a hard-failure.