add boot.util/dosh-timed
The regular `dosh` will block indefinitely if the called process does.
This can lead to tricky-to-debug behavior.

In many cases where `dosh` is used the commands are short-lived and could
be wrapped in a 1 or 2 second timeout, greatly improving the debugging
experience if commands happen to take longer than they should.