New task: call (#597)
This task can be used for ad-hoc code, either from the command line or
from build.boot. Use cases:

- run the project's entry point, like `lein run`
- experiments from the CLI: `boot call --print -f clojure-version`
- start a web server: `boot watch cljs call --once -f my.project/start`

As used from build.boot, the call task has the advantage that new users
don't need to learn about with-pass-thru to execute project code.

The `--eval` option allows to execute an arbitrary Clojure from. It is
similar to boot's `--init` but can be run as part of a boot pipeline.