add warning about improper use of repl :eval option
When users pass a string to the repl's `eval` option on the CLI this works fine:

    boot repl --eval "(println \"foo\")"

if they do the same in their build.boot or other Clojure code
the direct translation won't work:

    (repl :eval "(println \"foo\")")

The issue here is that the passed value is interpreted as edn and a string is a valid edn
value. This commit introduces a warning whenever the users passes a string to the :eval
option from Clojure code.

This issue probably affects other tasks making me think that there should be some kind
of `form` type for task options which are expected to be executable Clojure code.